Change short to float (for image type)

+++ Utilities/btkRescaleIntensity.cxx
@@ -72,7 +72,7 @@ int main(int argc, char *argv[])
     
   
   //ITK declaration
-  typedef short PixelType;
+  typedef float PixelType;
   const   unsigned int        Dimension = 3;
   typedef itk::Image< PixelType, Dimension >    ImageType;
   typedef itk::ImageRegionIterator< ImageType > itkIterator;
